requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
networkx>=3.2
python-multipart>=0.0.6
